% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docr.R
\name{anova_multi_formula}
\alias{anova_multi_formula}
\title{ANOVA modeling for variance partitioning}
\usage{
anova_multi_formula(
  data_use,
  feature_id_test,
  forms_to_test,
  value_var = "norm_abundance",
  n_min = 100,
  model_type = "simple",
  per_exp_var = "Sum Sq"
)
}
\arguments{
\item{data_use}{a dataframe with \code{value_var} numeric values
corresponding to the feature identified by \code{feature_id_test},
containing columns associated with covariates in \code{forms_to_test}}

\item{feature_id_test}{name of feature being tested. This is not a string
for filtering, but is used for print messages during model running.
Helpful when model fitting features in parallel}

\item{forms_to_test}{formulates to test, as character vector. Only
include covariates to the right of the tilde}

\item{value_var}{column names of dependent values, as a string. Defaults to
\code{norm_abundance} and is pasted to the left of the tilde in the formula}

\item{n_min}{minimum number of unique samples required to run ANOVA}

\item{model_type}{type of model to run, one of \code{"mixed"} for
\code{lmer()} fit, or \code{"simple"} for \code{lm()} fit}

\item{per_exp_var}{variable on which to calculate the percent of variance
explained, one of \code{"Sum Sq"} for net sum of squares, or
\code{"Mean Sq"} for mean sum of squares for each variable (sum of squares
divided by the degrees of freedom)}
}
\value{
a dataframe containing the model terms and percent variance
explained from \code{anova()}
}
\description{
This function runs \code{anova()} on one or more formulas to determine the
percent variance explained by each variable.
}
