% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{check_metadata}
\alias{check_metadata}
\title{Check metadata}
\usage{
check_metadata(metadata_tbl)
}
\arguments{
\item{metadata_tbl}{dataframe of metadata}
}
\value{
nothing, only checks
}
\description{
Check metadata
}
