% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docr.R
\name{coxph_model}
\alias{coxph_model}
\title{CoxPH Modeling}
\usage{
coxph_model(
  data_use,
  form,
  risk_var,
  event_var,
  value_var = "norm_abundance",
  n_min = 100
)
}
\arguments{
\item{data_use}{dataframe}

\item{form}{right-hand side of the formula with covariates for CoxPH}

\item{risk_var}{column name of risk data (i.e. lifespan), as a string}

\item{event_var}{column name of risk data censorship column, as a string.
Must be boolean or logical}

\item{value_var}{column names of metabolite values, as a string. Defaults to
\code{norm_abundance}}

\item{n_min}{minimum number of unique samples required to run coxph.
Set to low number to avoid model skipping on low n datasets}
}
\value{
dataframe of tidy coefficients from coxph modeling
}
\description{
This function performs CoxPH modeling using the survival::coxph package and
return a tidy summary of coefficients
}
