% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{cytoscape_clean_node_table}
\alias{cytoscape_clean_node_table}
\title{Clean a Cytoscape table}
\usage{
cytoscape_clean_node_table(network_name = NULL)
}
\arguments{
\item{network_name}{name of the Cytoscape network}
}
\value{
Cytoscape network with the specified table cleaned
}
\description{
Clean a Cytoscape table
}
