% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{cytoscape_mutate_network}
\alias{cytoscape_mutate_network}
\title{Reset Cytoscape network layout to default}
\usage{
cytoscape_mutate_network(network_name = NULL)
}
\arguments{
\item{network_name}{name of the network in Cytoscape}
}
\value{
reset network layout to default
}
\description{
Apply default changes to KEGG pathways in Cytoscape.
Since KGML files for KEGG pathways have the same format and properties,
same visual modifications can be applied to all KEGG networks in Cytoscape.
}
