% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{cytoscape_plot_node_color}
\alias{cytoscape_plot_node_color}
\title{Node color visualization}
\usage{
cytoscape_plot_node_color(
  data_input,
  id_mapping,
  metric,
  network_name = NULL,
  metric_range = 4
)
}
\arguments{
\item{data_input}{df, merged Cytoscape table of nodes with experimental data
must contain a "feature_name" column}

\item{id_mapping}{df that maps feature names to kegg ids
must contain "feature_name", "short_name", "id" columns}

\item{metric}{name of the metric to visualize}

\item{network_name}{name of the network in Cytoscape}

\item{metric_range}{range of data for plotting}
}
\value{
Cytoscape network with node color visualizing the metric of interest
}
\description{
Visualizing a metric with node color.
This functions works well for a comparison of two conditions
e.g. log2(treatment/control).
}
