% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{fdr}
\alias{fdr}
\title{Calculate qvalues}
\usage{
fdr(term_data, pval_var = "p.value")
}
\arguments{
\item{term_data}{dataframe of linear regression containing p.value}

\item{pval_var}{column name of pvalues in \code{term_data}, as a string.
Defaults to \code{"p.value"}}
}
\value{
dataframe of linear regression with qvalues added
}
\description{
Calculate qvalues
}
