% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{fdr_multi}
\alias{fdr_multi}
\title{Calculate q-values on multi-model or multi-term dataframes}
\usage{
fdr_multi(term_data, pval_var = "pvalue", nest_vars = NULL, padj_var = "padj")
}
\arguments{
\item{term_data}{dataframe containing pvalues columns requiring
adjustment}

\item{pval_var}{column name of pvalues to adjust, as a string}

\item{nest_vars}{column name(s) of nesting variables, if there are
multiple models, model terms, outcomes, etc. contained in \code{term_data}}

\item{padj_var}{column name for column with adjusted pvalues}
}
\value{
dataframe of linear regression with qvalues added
}
\description{
Calculate q-values on multi-model or multi-term dataframes
}
