% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingness.R
\name{generate_complete_dataset}
\alias{generate_complete_dataset}
\title{Generate complete dataset upon imputing missing values}
\usage{
generate_complete_dataset(mzroll_list, metadata, lod_values)
}
\arguments{
\item{mzroll_list}{data in triple omic structure that's processed
with claman::process_mzroll
samples table of mzroll_list must contain a numeric InjOrder column}

\item{metadata}{metadata of experiment}

\item{lod_values}{a tibble that maps groupId
to log2 feature-level imputation values}
}
\value{
dataframe of complete dataset
}
\description{
Generate complete dataset upon imputing missing values
}
