% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{import_metadata}
\alias{import_metadata}
\title{Import metadata from a Google Sheet}
\usage{
import_metadata(sheet_id)
}
\arguments{
\item{sheet_id}{Google Sheet id}
}
\value{
a dataframe of metadata
}
\description{
Import metadata from a Google Sheet
}
