% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingness.R
\name{imputed_comparisons}
\alias{imputed_comparisons}
\title{Find comparisons that have at least one imputed value}
\usage{
imputed_comparisons(feature_id, cond, metadata, df)
}
\arguments{
\item{feature_id}{groupId from mzrolldb}

\item{cond}{Condition in metadata}

\item{metadata}{metadata of experiment}

\item{df}{dataframe in long format}
}
\value{
dataframe of terms and groupIds for comparisons with imputed values
}
\description{
Find comparisons that have at least one imputed value
}
