% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{lm}
\alias{lm}
\title{Pairwise linear regression for one feature and one condition}
\usage{
lm(feature_id, ref_cond_num, metadata, df)
}
\arguments{
\item{feature_id:}{groupId from mzrolldb}

\item{cond_num:}{ConditionNum in metadata}

\item{metadata:}{metadata of experiment}

\item{df:}{dataframe in long format}
}
\value{
one-row data from linear regression
}
\description{
Pairwise linear regression for one feature and one condition
}
\examples{
complete_dataset <- generate_complete_dataset(mzrolldb_file_path, metadata)
list_complete_dataset <- purrr::map(1:5, ~generate_complete_dataset(mzrolldb_file_path, metadata))

}
