% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{lm_feature}
\alias{lm_feature}
\title{Pairwise linear regression}
\usage{
lm_feature(feature_id, ref_cond, metadata, df, quant_var = "log2_abundance")
}
\arguments{
\item{feature_id}{groupId from mzrolldb}

\item{ref_cond}{RefCondition in metadata}

\item{metadata}{metadata of experiment}

\item{df}{dataframe in long format}

\item{quant_var}{string of quant variable to be used for lm}
}
\value{
one-row data from linear regression
}
\description{
This function will perform one-feature regression
for all conditions that have the same reference condition
}
