% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{lm_pool}
\alias{lm_pool}
\title{Pooled results for linear regression}
\usage{
lm_pool(feature_id, conds, metadata, df_list, quant_var = "log2_abundance")
}
\arguments{
\item{feature_id}{groupId(s) from mzrolldb}

\item{conds}{Condition(s) in metadata}

\item{metadata}{metadata of experiment}

\item{df_list}{list of complete dataframes upon imputing missing peaks}

\item{quant_var}{string of quant variable to be used for lm}
}
\value{
df of linear regression summary
for one feature and all pairwise comparisons
}
\description{
pooling regression results from
multiple imputation of missing peaks
}
