% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docr.R
\name{lmer_multi_formula}
\alias{lmer_multi_formula}
\title{Linear mixed modeling}
\usage{
lmer_multi_formula(
  data_use,
  feature_id_test,
  forms_to_test,
  value_var = "norm_abundance",
  n_samples = NULL,
  n_sample_threshold = 0.1,
  model_type = "mixed"
)
}
\arguments{
\item{data_use}{a dataframe with \code{value_var} numeric values
corresponding to the feature identified by \code{feature_id_test},
containing columns associated with covariates in \code{forms_to_test}}

\item{feature_id_test}{name of feature being tested. This is not a string
for filtering, but is used for print messages during model running. Helpful
when model fitting features in parallel}

\item{forms_to_test}{formulates to test, as character vector. Only include
covariates to the right of the tilde}

\item{value_var}{column names of dependent values, as a string. Defaults to
\code{norm_abundance} and is pasted to the left of the tilde in the formula}

\item{n_samples}{number of unique samples in experiment. Defaults to
\code{NULL} to ignore parameter.}

\item{n_sample_threshold}{percent of \code{n_samples} below which model
fitting will skip. Defaults to 0.1 for 10\%.}

\item{model_type}{type of model to run, one of \code{"mixed"} for
\code{lmer()} fit, or \code{"simple"} for \code{lm()} fit}
}
\value{
a dataframe containing the feature_id, model_term, model, coef,
se_coef, p_value_coef, t_value, and AIC of the model fit
}
\description{
This function trains a linear mixed model through \code{lmer} as specified
}
