% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docr.R
\name{metstats_elastic_data_parse}
\alias{metstats_elastic_data_parse}
\title{Test and Train Data Parsing}
\usage{
metstats_elastic_data_parse(
  df,
  dependent_var,
  sample_id_var,
  dev_ids,
  val_ids,
  key_var = "feature_id",
  value_var = "norm_abundance",
  key_var_filter = NULL,
  makeX_na_impute = TRUE
)
}
\arguments{
\item{df}{a dataframe containing columns described below}

\item{dependent_var}{column name of regression outcome varibale,
as a string}

\item{sample_id_var}{column name of observation identifiers, as a string,
on which \code{dev_ids} and \code{val_ids} will be parsed}

\item{dev_ids}{character vector corresponding to values in
\code{sample_id_var} on which to filter observations for model development}

\item{val_ids}{character vector corresponding to values in
\code{sample_id_var} on which to filter observations for model validation}

\item{key_var}{column name of identifier variables to be included on the
right hand side of the model equation, as a string; defaults to "feature_id"}

\item{value_var}{column name of values associated with \code{value_var} to
be included on the right hand side of the equation, as a string; defaults to
"norm_abundance"}

\item{key_var_filter}{character string on which variables associated with
\code{key_var} can be filtered out with grep; i.e. all unknown compounds
will be removed when default \code{NULL} is changed to "unk"}

\item{makeX_na_impute}{logical to indicate if the \code{glmnet::makeX}
function should handle missing values (\code{metstats::metstats_elastic_net}
will not work with missing values)}
}
\value{
a list containing dataframes x.dev, y.dev, x.val, y.val for direct
input into \code{metstats::metstats_elastic_net}
}
\description{
Given a dependent variable, development IDs, and validation IDs,
this function returns a list with 4 dataframes corresponding to development
and validation dataframes for x and y variables. This function utilizes
\code{glmnet} and is designed to work with
\code{metstats::metstats_elastic_net} using the development dataframes in
the return list
}
