% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docr.R
\name{metstats_elastic_net}
\alias{metstats_elastic_net}
\title{Elastic Net Regularization Regression}
\usage{
metstats_elastic_net(
  x,
  y,
  vars_not_penalized = NULL,
  its = 100,
  train_percent = 0.75,
  alpha_levels = 5,
  key_var = "feature_id",
  makeX_na_impute = FALSE
)
}
\arguments{
\item{x}{a dataframe with observations on rows and columns as model
ovariates}

\item{y}{a double or matrix of outcome values with observations
corresponding to those in \code{x}}

\item{vars_not_penalized}{character vector of column names in \code{x} that
will recieve no penalty in \code{glmnet::cv.glmnet}}

\item{its}{number of iterations to run the model; defaults to 100}

\item{train_percent}{percent of dataset provided on which to determine the
appropriate alpha based on lowest MSE; defaults to 0.75}

\item{alpha_levels}{number of alpha levels to test which will be divided
between 0 and 1 proportionally; defaults to (defaults to 5 which corresponds
to alpha = 0, 0.25, 0.5, 0.75, and 1)}

\item{key_var}{column name of identifier variables, as a string;
defaults to "feature_id". Used for naming the column in the return dataframe
and does not impact modeling}

\item{makeX_na_impute}{logical to indicate if the \code{glmnet::makeX}
function should handle categorical values (convert to binaries) and impute
means on missing values}
}
\value{
a list containing 2 dataframes: \code{Selection} which contains all
iteration summary data, and \code{Models} which returns all stored model fits
}
\description{
This function trains an elastic net model for multiple alpha levels.
There are options to not penalize specific variables (as grepped column
names), change the number of iterations (default \code{its = 100}), and change
the percent of development dataset that is used for training versus
testing for alpha value (default \code{train_percent = 0.75}). This function
runs \code{glmnet::cv.glmnet}.
}
