% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msea.R
\name{pathway_enrichment}
\alias{pathway_enrichment}
\title{Pathway Enrichment Analysis}
\usage{
pathway_enrichment(
  data_input,
  id_column,
  feature_sets_file,
  ranking_metric,
  metric_type = "non-abs",
  pval = 1
)
}
\arguments{
\item{data_input}{a dataframe, must contain an id column and
a ranking metric column
if an optional 'feature_name' column exists in the input data,
' id_column' will be replaced by 'feature_name' in the final enrichment table}

\item{id_column}{name of the column to be used as id}

\item{feature_sets_file}{path to the feature sets file with a
standard GSEA tab-delimited text file format, see the link below:
https://software.broadinstitute.org/cancer/software/gsea/wiki/index.php/Data_formats
for metabolite enrichment analysis with KEGG's pathways
this file is embedded within the package:
"/metstats/inst/kegg_msea.txt"}

\item{ranking_metric}{column name from input data
to be used as the metric for enrichment analysis}

\item{metric_type}{type of the metric being used
for enrichment analysis 'abs' or 'non-abs'}

\item{pval}{pvalue threshold to filter the enrichment results}
}
\value{
a list that contains
\enumerate{
\item a tbl of enrichment statistics
\item an enrichment dot plot
\item a list of mountain plots
}
}
\description{
To perform Pathway Enrichment Analysis with any omics data type
}
