% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{PDB_compounds}
\alias{PDB_compounds}
\title{Table of peakgroupmatch from file}
\usage{
PDB_compounds(mzrolldb_file_path)
}
\arguments{
\item{mzrolldb_file_path:}{file path to mzrolldb file}
}
\value{
compounds from mzrolldb file
}
\description{
Table of peakgroupmatch from file
}
