% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{qqq_lm_pool}
\alias{qqq_lm_pool}
\title{Pooled results for linear regression of one feature and all pairwise comparisons upon multiple imputation of missing peaks}
\usage{
qqq_lm_pool(feature_id, conds, metadata, df_list, quant_var = "log2_abundance")
}
\arguments{
\item{feature_id}{groupId(s) from mzrolldb}

\item{conds}{Condition(s) in metadata}

\item{metadata}{metadata of experiment}

\item{df_list}{list of complete dataframes upon imputing missing peaks}

\item{quant_var}{string of quant variable to be used for lm}
}
\value{
dataframe of linear regression summary for one feature and all pairwise comparisons
}
\description{
Pooled results for linear regression of one feature and all pairwise comparisons upon multiple imputation of missing peaks
}
