% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ST003519_id_mapping}
\alias{ST003519_id_mapping}
\title{id mapping}
\format{
A data frame:
\describe{
\item{feature_name}{feature names in the dataset}
\item{id}{KEGG id for features}
\item{short_name}{short name for fatures}
}
}
\usage{
ST003519_id_mapping
}
\description{
KEGG id mapping for the features of Metabolomics Workbench ST003519 dataset
}
\seealso{
Other ST003519: 
\code{\link{ST003519_metadata}},
\code{\link{ST003519_mzroll}()}
}
\concept{ST003519}
\keyword{datasets}
