% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ST003519_metadata}
\alias{ST003519_metadata}
\title{metadata}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 5 columns.
}
\usage{
ST003519_metadata
}
\description{
metadata example for Metabolomics Workbench ST003519 dataset
\describe{
\item{SampleName}{sub string of sample filename, must be a unique match for one sample}
\item{Condition}{condition that the samle belongs to}
\item{RefCondition}{reference condition for the condition that the samle belongs to}
\item{Replicate}{replicate number for the sample}
\item{treatment}{example of a variable, other dataset variables must be included as additional columns}
}
}
\seealso{
Other ST003519: 
\code{\link{ST003519_id_mapping}},
\code{\link{ST003519_mzroll}()}
}
\concept{ST003519}
\keyword{datasets}
