% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ST003519_mzroll}
\alias{ST003519_mzroll}
\title{mzrollDB}
\usage{
ST003519_mzroll()
}
\value{
path to the .mzrollDB dataset
}
\description{
mzrollDB file for Metabolomics Workbench ST003519 dataset
}
\details{
The experiment contains 3 experimental conditions
\itemize{
\item{control, ischemia, insuline_ischemia}
}
}
\seealso{
Other ST003519: 
\code{\link{ST003519_id_mapping}},
\code{\link{ST003519_metadata}}
}
\concept{ST003519}
