% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Volcano plot}
\usage{
volcano_plot(lm_data, fdr_cutoff = 0.05, feature_labels = NULL)
}
\arguments{
\item{lm_data}{dataframe of regression analysis}

\item{fdr_cutoff}{FDR cutoff to label for significance}

\item{feature_labels}{list of compound names from
compoundName column to label on the plot}
}
\value{
a grob
}
\description{
Volcano plot
}
