% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{fdr}
\alias{fdr}
\title{Find qvalues}
\usage{
fdr(term_data)
}
\arguments{
\item{term_data:}{dataframe of linear regression conataining p.value}
}
\value{
dataframe of linear regression with qvalues added
}
\description{
Find qvalues
}
\examples{
lm_fdr <- fdr(lm_data)

}
