% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missignness.R
\name{generate_complete_dataset}
\alias{generate_complete_dataset}
\title{Generate complete dataset upon imputing missing peaks}
\usage{
generate_complete_dataset(mzrolldb_file_path, metadata, lod_values)
}
\arguments{
\item{mzrolldb_file_path:}{mzrolldb file path}

\item{metadata:}{metadata of experiment}
}
\value{
dataframe of complete dataset
}
\description{
Generate complete dataset upon imputing missing peaks
}
\examples{
complete_dataset <- generate_complete_dataset(mzrolldb_file_path, metadata)
list_complete_dataset <- purrr::map(1:5, ~generate_complete_dataset(mzrolldb_file_path, metadata))

}
