% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{imputed_comparisons}
\alias{imputed_comparisons}
\title{Find comparisons that have at least one imputed value}
\usage{
imputed_comparisons(feature_id, cond_num, metadata, df)
}
\arguments{
\item{feature_id:}{groupId from mzrolldb}

\item{cond_num:}{ConditionNum in metadata}

\item{metadata:}{metadata of experiment}

\item{df:}{dataframe in long format}
}
\value{
dataframe of terms and groupIds for comparisons with imputed values
}
\description{
Find comparisons that have at least one imputed value
}
\examples{
df_imputed <- imputed_comparisons(2, 3, metadata, df)

}
