% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missignness.R
\name{imputed_comparisons}
\alias{imputed_comparisons}
\title{Find comparisons that have at least one imputed value}
\usage{
imputed_comparisons(feature_id, cond_num, metadata, df)
}
\arguments{
\item{feature_id:}{groupId for specific feature}

\item{cond_num:}{ConditionNum in metadata}

\item{metadata:}{df of metadata}

\item{df:}{data in long format}
}
\description{
Find comparisons that have at least one imputed value
}
