% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{lm_multi}
\alias{lm_multi}
\title{Linear regression for multiple conditions}
\usage{
lm_multi(feature_id, condition_nums, metadata, df)
}
\arguments{
\item{feature_id:}{groupId(s) from mzrolldb}

\item{condition_nums:}{ConditionNum(s) in metadata}

\item{metadata:}{metadata of experiment}

\item{df:}{dataframe in long format}
}
\value{
one-row data from linear regression
}
\description{
Linear regression for multiple conditions
}
\examples{
lm_all <- purrr::map(feature_ids, ~lm_all(.x, ref_ids, metadata, complete_data))

}
