% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{lm_pool}
\alias{lm_pool}
\title{Pooled results for linear regression of one feature and all pairwise comparisons upon multiple imputation of missing peaks}
\usage{
lm_pool(feature_id, ref_condition_nums, metadata, df_list)
}
\arguments{
\item{feature_id:}{groupId(s) from mzrolldb}

\item{ref_condition_nums:}{RefConditionNum(s) in metadata}

\item{metadata:}{metadata of experiment}

\item{df_list:}{list of complete dataframes upon imputing missing peaks}
}
\value{
dataframe of linear regression summary for one feature and all pairwise comparisons
}
\description{
Pooled results for linear regression of one feature and all pairwise comparisons upon multiple imputation of missing peaks
}
\examples{
lm_pooled <- lm_pool(2, c(1,2,3), metadata, df_list)

}
